Analyzer:  - Started working on parsing to cosine differences
LargeTextProcessor:
 - Fixed bug where the user would leave no review text and the parser wouldn't catch it

+++ LargeTextProcessor.py
@@ -2,19 +2,10 @@ import time
 import numpy as np
 import sys
 from scipy.sparse import *
-from numpy import int8
+from numpy import int8, int
 from scipy import float16
 import pickle
 
-class User:
-    '''
-    Contains profile information about a user.
-    '''
-
-    def __init__(self, ID):
-        self.ID = int(ID)
-    
-    
 class Beer:
     '''
         Name: Name of the beer.... duh
@@ -48,12 +39,15 @@ class Review:
         self.text = text
 
 
+beersToProcess = 1586300             #How many reviews to we want to process
+overwritePickle = True              #Overwrite the pickle output file with this run?
+
 '''
 Do line-by-line read of the file (won't load it all into memory) and store
 in above class structures.
 '''
 #CHANGE FILEPATH HERE FOR YOUR MACHINE
-with open("beeradvocate.txt") as infile:
+with open("/media/eljefe/BC7A0ADA7A0A9176/beeradvocate.txt") as infile:
     #Beer Info
     name = "undefined"
     beerID = -1
@@ -74,21 +68,59 @@ with open("beeradvocate.txt") as infile:
     profileName = "undefined"
     
     reviewList = []
-    beersToProcess = 1750000
     counter = 0
     
     #Create a unique user ID starting at 0
-    originalUserIDtoNewUserIDMap = {}
+    originalUsernameToNewUserIDMap = {}
     userIDCounter = 0           #For unique user ID
     
     #Create a unique beer ID starting at 0
     originalBeerIDtoNewBeerIDMap = {}
     beerIDCounter = 0           #For unique beer ID
     
+    print '[INFO] Beginning to parse reviews'
+    if not overwritePickle:
+        print '[WARNING] This is a test run only! The output will NOT be pickled.'
+        
     startTime = time.clock()                 #Time for profiling
     for rawLine in infile:                   #Iterate through every line, one at a time
         if counter >= beersToProcess:        #Only check subset of beers                                        
             break
+        
+        if rawLine.isspace():       #Space separator
+            #Create the review
+            review = Review(profileName, ratedBeer, aroma, palate, taste, overall, reviewTime, text)
+            reviewList.append(review)
+            counter += 1
+            
+          #Reset parser info
+            #Beer Info
+            name = ""
+            beerID = -1
+            brewerID = -1
+            abv = 0
+            style = "undefined"
+            
+            #Review Info
+            appearance = -1
+            aroma = -1
+            palate = -1 
+            taste = -1
+            overall = -1
+            reviewTime = 0
+            text = "undefined"
+            
+            #User
+            profileName = "undefined"
+            
+            if counter % int(0.01*beersToProcess) == 0:
+                bar_length = 100
+                percent = float(counter) / beersToProcess
+                hashes = '#' * int(round(percent * bar_length))
+                spaces = ' ' * (bar_length - len(hashes))
+                sys.stdout.write("\rProgress: [{0}] {1}%".format(hashes + spaces, int(round(percent * 100))))
+                sys.stdout.flush()
+        
         line = rawLine.rstrip()              #Remove newline & any unnecessary white space
         
         if 'beer/name: ' in line:
@@ -125,48 +157,24 @@ with open("beeradvocate.txt") as infile:
             reviewTime = line.replace('review/time: ', '')
         elif 'review/profileName:' in line:
             profileName = line.replace('review/profileName: ', '')
-            if profileName in originalUserIDtoNewUserIDMap:          #Check if we have already created unique profile ID for this username
-                profileName = originalUserIDtoNewUserIDMap[profileName]
+            if profileName in originalUsernameToNewUserIDMap:          #Check if we have already created unique profile ID for this username
+                profileName = originalUsernameToNewUserIDMap[profileName]   #Use this mapping for the new profile name
             else:                                                    #Create new mapping of real user ID to next unique ID
-                originalUserIDtoNewUserIDMap[profileName] = userIDCounter
+                originalUsernameToNewUserIDMap[profileName] = userIDCounter
                 profileName = userIDCounter             #Replace real username with user ID 
                 userIDCounter += 1
         
         elif 'review/text: ' in line:
 #             text = line.replace('review/text: ', '')            
             text = ''           #Toss out the review text for sake of memory
-            review = Review(profileName, ratedBeer, aroma, palate, taste, overall, reviewTime, text)
-            reviewList.append(review)
-            counter += 1
             
-          #Reset parser info
-            #Beer Info
-            name = ""
-            beerID = -1
-            brewerID = -1
-            abv = 0
-            style = "undefined"
-            
-            #Review Info
-            appearance = -1
-            aroma = -1
-            palate = -1 
-            taste = -1
-            overall = -1
-            reviewTime = 0
-            text = "undefined"
-            
-            #User
-            profileName = "undefined"
-             
-            if counter % 1000 == 0:
-                print 'Processed ' + str(counter) + ' reviews'
+               
                 
 #Display parse performance
 reviewTimeToProcess = time.clock() - startTime
-print '[INFO]: Processed ' + str(len(reviewList)) + ' beer reviews in ' + str(reviewTimeToProcess) + ' seconds'
-print 'Total of: ' + str(sys.getsizeof(reviewList)) + ' bytes'
-print
+print '\n[INFO]: Processed ' + str(len(reviewList)) + ' beer reviews in ' + str(reviewTimeToProcess) + ' seconds'
+print '[INFO] Total of: ' + str(sys.getsizeof(reviewList)) + ' bytes'
+print '-----------------------------'
 #Update users to have unique, sequential IDs
 
 #User-review mapping
@@ -200,37 +208,61 @@ for review in reviewList:
 beerList = []
 for i in range(0,beerIDCounter):
     beerList.append(i)
+    
+if(len(beerList) != len(set(beerList))):
+    print '[ERROR] Duplicate beers found in the beer list!'
 
 #Generate list of users (straight from keys)
 userList = []
 for user in userReviewMap:
     userList.append(user)
     
-#Convert into format of row,column,data coordinates
-row = []            #Users
-column = []         #Beers
-data = []           #Reviews
+lastUser = 0
+for i in range(1,len(userList)-1):
+    user = userList[i]
+    if user - lastUser != 1:
+        print user
+    lastUser = user
 
-for user in userReviewMap:      #A single row
+if(len(userList) != len(set(userList))):
+    print '[ERROR] Duplicate users found in the user list!'
+    
+#Convert into format of row,column,data coordinates (see: https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.csr_matrix.html#scipy.sparse.csr_matrix)
+row = []            #Users (this is the Y coordinate of where data goes in the matrix)
+column = []         #Beers (this is the X coordinate of where data goes in the matrix)
+data = []           #Reviews (this is the data in the matrix)
+
+for user in userList:      #A single row
     for review in userReviewMap[user]:
         row.append(user)            #User ID as row
         column.append(review.beer.ID)
         data.append(review.overall)
+print '[DEBUG] Last mapping: (' + str(row[len(row)-1]) + ', ' + str(column[len(column)-1]) + ')'
+
+#Check if user has at least one review:
+for user in userReviewMap:
+    if len(userReviewMap[user]) == 0:
+        print '[ERROR] A user was pulled from the database with no reviews. What happened?! This will break the analyzer with divide by zero problems'
 
 #Convert all to numpy arrays
 numPiRow = np.array(row)
 numPiColumn = np.array(column)
 numpiData = np.array(data)
 
+#Create sparse array (compressed by row)
+print '[INFO] Creating sparse matrix with dimensions (' + str(len(userList)) + ', ' + str(len(beerList)) + ')'
+userBeerReviewArray = csr_matrix((numpiData, (numPiRow, numPiColumn)), shape=(len(userList),len(beerList)), dtype=int)
 
-#Create sparse array (compressed 
-userBeerReviewArray = csr_matrix((data, (row,column)), shape=(len(column),len(row)), dtype=int)
-
-dumpfile = open("beerArray.pickle", 'w')
-pickle.dump(userBeerReviewArray, dumpfile)
-
-
-
-
-
+#Dump out to file
+if overwritePickle:
+    print '[INFO] Pickling sparse matrix out to beerArray.pickle'
+    try:
+        dumpfile = open("beerArray.pickle", 'w')
+        pickle.dump(userBeerReviewArray, dumpfile)
+        print '[INFO] Finished pickling sparse matrix'
+    except Exception as e:
+        print '[ERROR] Pickling operation failed!: ' + str(e)
+else:
+    print '[INFO] Pickle overwrite disabled. Not writing out to filesystem'
+
